+++ Packages/com.seahyx.xruitoolkit/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.4.0
+
+## Added
+
+- Added new `ImageColourVisualEffect`
+- Added new `SpriteColourVisualEffect`
+
+## Changed
+
+- Changed all interpolation methods to unclamped
+- Renamed `ColourVisualEffect` to `MaterialColourVisualEffect`
+
 ## 1.3.0
 
 ### Added